add an optional hasher extension
