+++ handle_misc.go
@@ -2,19 +2,11 @@
 package ftpserver
 
 import (
-	"bufio"
-	"crypto/md5"  //nolint:gosec
-	"crypto/sha1" //nolint:gosec
-	"crypto/sha256"
-	"crypto/sha512"
+	"bufio" //nolint:gosec
+	//nolint:gosec
 	"crypto/tls"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"hash"
-	"hash/crc32"
-	"io"
-	"os"
 	"strings"
 	"time"
 )
@@ -238,51 +230,4 @@ func (c *clientHandler) handleQUIT() error {
 	c.reader = nil
 
 	return nil
-}
-
-func (c *clientHandler) computeHashForFile(filePath string, algo HASHAlgo, start, end int64) (string, error) {
-	var h hash.Hash
-	var file FileTransfer
-	var err error
-
-	switch algo {
-	case HASHAlgoCRC32:
-		h = crc32.NewIEEE()
-	case HASHAlgoMD5:
-		h = md5.New() //nolint:gosec
-	case HASHAlgoSHA1:
-		h = sha1.New() //nolint:gosec
-	case HASHAlgoSHA256:
-		h = sha256.New()
-	case HASHAlgoSHA512:
-		h = sha512.New()
-	default:
-		return "", errUnknowHash
-	}
-
-	if fileTransfer, ok := c.driver.(ClientDriverExtentionFileTransfer); ok {
-		file, err = fileTransfer.GetHandle(filePath, os.O_RDONLY, start)
-	} else {
-		file, err = c.driver.OpenFile(filePath, os.O_RDONLY, os.ModePerm)
-	}
-
-	if err != nil {
-		return "", err
-	}
-
-	if start > 0 {
-		_, err = file.Seek(start, io.SeekStart)
-		if err != nil {
-			return "", err
-		}
-	}
-
-	_, err = io.CopyN(h, file, end-start)
-	defer file.Close() //nolint:errcheck // we ignore close error here
-
-	if err != nil && err != io.EOF {
-		return "", err
-	}
-
-	return hex.EncodeToString(h.Sum(nil)), nil
 }
